x11: Implement run_return using calloop
